Make sure separators are all the same.
If a -sep pattern allows for different separators, we still want to
enforce that all separators in the same number are the same. In order
to do so, there will be mandatory capture, that is, even without -Keep
there will be capture. The name of the capture will be different for
each pattern.